completr with stretch

+++ src/components/ShoppingList/ShoppingList.jsx
@@ -1,21 +1,11 @@
-import { useState, useEffect } from 'react';
 import DisplayItem from '../DisplayItem/DisplayItem';
 import axios from 'axios';
+import './ShoppingList.css';
 
 function ShoppingList (props) {
-    // const [shoppingList, setShoppingList] = useState([]);
-
-    // const getShoppingList = () => {
-    //     axios.get('/shoppinglist').then((response) => {
-    //         console.log(response.data);
-    //         setShoppingList(response.data)
-    //     }).catch((error) => {
-    //         console.log('GET/shoppinglist error', error);
-    //         alert('Something went wrong getting your list!');
-    //     });
-    // };
 
     const resetList = () => {
+        if (window.confirm('Are you sure you want to remove all purchases? This cannot be undone.')) {
         axios.put('/shoppinglist/all').then((response) =>{
             console.log("All items reset as unpurchased.", response);
             props.getShoppingList();
@@ -24,12 +14,13 @@ function ShoppingList (props) {
             console.error("Error in PUT '/shoppinglist/all' inside ShoppingList component.", error);
             alert("Error in PUT '/shoppinglist/all' inside ShoppingList component. See console.");
         });
+    } else {
+        return;
     }
-    // useEffect(() => {
-    //     getShoppingList();
-    // }, []);
+    }
 
     const clearList = () => {
+      if (window.confirm('Are you sure you want to delete your list? This cannot be undone.')) {
         axios.delete('/shoppinglist/all').then((response) => {
             console.log("List cleared successfully", response);
             props.getShoppingList();
@@ -38,15 +29,19 @@ function ShoppingList (props) {
             console.error("Error in DELETE '/shoppinglist/all' inside ShoppingList component.", error);
             alert("Error in DELETE '/shoppinglist/all' inside ShoppingList component. See console.");
         });
+    } else {
+        return;
     }
-
+    }
 
     return (
         <div id="shopping-list">
-            <button onClick={resetList}>Reset</button>
-            <button onClick={clearList}>Clear</button>
-            <div>
-
+            <hr></hr>
+            <h3>Shopping List:</h3>
+            <button onClick={resetList}>Reuse List</button>
+            <button onClick={clearList}>Delete All</button>
+            <hr></hr>
+            <div id="shopping-list-body">
                 {props.list.map((item) => (<DisplayItem getShoppingList={props.getShoppingList} key={item.id} item={item}/>))}
             </div>
         </div>
